outpass_project/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
mysql-connector-python==8.2.0
pydantic==2.5.3
